Modify the build's handling of the Windows HID libraries.
Add alternate CMake modules for finding the HID headers/libraries.

Since we can now build with newer than DDK 7600, there is a define that
has taken place of the pure MSVC version check in two headers. Build with
newer Windows SDKs and compilers, and its status as off by default will be
right.

(This also fixes the build of libfreespace with msys2/mingw-w64)